// apps/admin-web/src/.generated/api/model/adminCustomerItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

export interface AdminCustomerItem {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  image?: string | null;
  emailVerified: boolean;
  banned: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  roleCode?: string | null;
  /** @nullable */
  roleLabel?: string | null;
}